+++ oci/adb.go
@@ -1,0 +1,14 @@
+// Copyright (c) 2026, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package oci
+
+import (
+	"context"
+
+	adb "github.com/oracle/oci-go-sdk/v65/database"
+)
+
+func ListAllDatabases(ctx context.Context, client adb.DatabaseClient) {
+
+}